+++ dao/game_result_dao.go
@@ -2,17 +2,11 @@ package dao
 
 import (
 	"context"
-	"errors"
-	"github.com/ildomm/cceab/domain"
+	"github.com/google/uuid"
+	"github.com/ildomm/cceab/entity"
 )
 
-var ErrGameResultNotFound = errors.New("game result not found")
-var ErrTransactionIdExists = errors.New("transaction id already exists")
-var ErrInvalidGameStatus = errors.New("invalid game status")
-var ErrInvalidTransactionSource = errors.New("invalid transaction source")
-var ErrInvalidTransactionStatus = errors.New("invalid transaction status")
-
 type GameResultDAO interface {
-	CreateGameResult(ctx context.Context, gameResult domain.GameResult) (*domain.GameResult, error)
+	CreateGameResult(ctx context.Context, userId uuid.UUID, gameStatus entity.GameStatus, Amount float64, transactionID string) (*entity.GameResult, error)
 	ValidateGameResults(ctx context.Context, totalGamesToCancel int) error
 }