+++ database/querier.go
@@ -3,20 +3,21 @@ package database
 import (
 	"context"
 	"github.com/google/uuid"
-	"github.com/ildomm/cceab/domain"
+	"github.com/ildomm/cceab/entity"
 )
 
 type Querier interface {
 	Close()
 
-	InsertGameResult(ctx context.Context, gameResult domain.GameResult) error
+	InsertGameResult(ctx context.Context, gameResult entity.GameResult) error
 
-	SelectUser(ctx context.Context, userId uuid.UUID) (*domain.User, error)
-	SelectUsersByValidationStatus(ctx context.Context, validationStatus bool) ([]domain.User, error)
+	SelectUser(ctx context.Context, userId uuid.UUID) (*entity.User, error)
+	SelectUsersByValidationStatus(ctx context.Context, validationStatus bool) ([]entity.User, error)
 
-	SelectGameResultByTransaction(ctx context.Context, transactionId string) (*domain.GameResult, error)
-	SelectGameResultsByUser(ctx context.Context, userId uuid.UUID, validationStatus domain.ValidationStatus) ([]domain.GameResult, error)
+	GameResultExists(ctx context.Context, transactionId string) (bool, error)
+	// SelectGameResultByTransaction(ctx context.Context, transactionId string) (*entity.GameResult, error)
+	SelectGameResultsByUser(ctx context.Context, userId uuid.UUID, validationStatus entity.ValidationStatus) ([]entity.GameResult, error)
 
 	UpdateUserBalance(ctx context.Context, userId uuid.UUID, amount float64, validationStatus bool) error
-	UpdateGameResult(ctx context.Context, gameResultId uuid.UUID, validationStatus domain.ValidationStatus) error
+	UpdateGameResult(ctx context.Context, gameResultId uuid.UUID, validationStatus entity.ValidationStatus) error
 }